Small chnages to Scheme and moving tests
